Create paragraph component
